requirements.txt
flask
flask-login
pyjwt
pandas
werkzeug
openpyxl
SQLAlchemy>=2.0
alembic
flask-mailman
